demo/image-compression-deploy/requirements.txt
asgiref==3.7.1
Django==4.2.1
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
sqlparse==0.4.4
typing_extensions==4.6.2